Feat/#95

+++ src/main/java/com/peeroreum/service/ChallengeService.java
@@ -8,7 +8,6 @@ import com.peeroreum.domain.image.Image;
 import com.peeroreum.dto.member.MemberProfileDto;
 import com.peeroreum.dto.wedu.ChallengeMemberList;
 import com.peeroreum.dto.wedu.ChallengeReadDto;
-import com.peeroreum.dto.wedu.WeduMonthlyProgressDto;
 import com.peeroreum.repository.ChallengeImageRepository;
 import com.peeroreum.repository.MemberWeduRepository;
 import com.peeroreum.service.attachment.ImageService;
@@ -39,7 +38,7 @@ public class ChallengeService {
         LocalDate now = LocalDate.now();
         MemberWedu memberWedu = memberWeduRepository.findByMemberAndWedu(member, wedu);
         if(challengeImageRepository.existsByWeduAndMemberAndChallengeDate(wedu, member, now)) {
-            deleteChallengeImages(member, wedu);
+            deleteTodayChallengeImages(member, wedu);
             challengeImageRepository.deleteByWeduAndMemberAndChallengeDate(wedu, member, now);
         }
 
@@ -69,12 +68,14 @@ public class ChallengeService {
         return new ChallengeReadDto(imageUrls);
     }
 
-    public void deleteChallengeImages(Member member, Wedu wedu) {
-        ChallengeImage challengeImage = challengeImageRepository.findAllByMemberAndWeduAndChallengeDate(member, wedu, LocalDate.now());
+    public void deleteTodayChallengeImages(Member member, Wedu wedu) {
+        LocalDate now = LocalDate.now();
+        ChallengeImage challengeImage = challengeImageRepository.findAllByMemberAndWeduAndChallengeDate(member, wedu, now);
         List<Image> images = challengeImage.getImage();
         for(Image image : images) {
             imageService.deleteImage(image.getId());
         }
+        challengeImageRepository.deleteAllByMemberAndWeduAndChallengeDate(member, wedu, now);
     }
 
     public ChallengeMemberList readChallengeMembers(Wedu wedu, LocalDate formattedDate) {
@@ -109,8 +110,29 @@ public class ChallengeService {
         return progressList;
     }
 
-    public void deleteChallengeImages(Wedu wedu) {
-        challengeImageRepository.deleteAllByWedu(wedu);
+    public void deleteTodayChallengeImages(Wedu wedu) {
+        List<ChallengeImage> challengeImages = challengeImageRepository.findAllByWedu(wedu);
+        if(!challengeImages.isEmpty()) {
+            deleteChallengeImages(challengeImages);
+            challengeImageRepository.deleteAllByWedu(wedu);
+        }
+    }
+
+    public void deleteByWeduAndMember(Wedu wedu, Member member) {
+        List<ChallengeImage> challengeImages = challengeImageRepository.findAllByWeduAndMember(wedu, member);
+        if(!challengeImages.isEmpty()) {
+            deleteChallengeImages(challengeImages);
+            challengeImageRepository.deleteAllByWeduAndMember(wedu, member);
+        }
+    }
+
+    public void deleteChallengeImages(List<ChallengeImage> challengeImages) {
+        for(ChallengeImage challengeImage : challengeImages) {
+            List<Image> images = challengeImage.getImage();
+            for(Image image : images) {
+                imageService.deleteImage(image.getId());
+            }
+        }
     }
 
     public Long getTodayProgress(Wedu wedu, int total) {
